Merge Updates from NotAKidOnSteam (#1)
* Fixed SetProxyColor() on Client  
Leftover code was replaced to fix the client SetProxyColor() function. If called on the client it will set locally only.  
Very useful for model previews in clientside menus.  

* Grammer Fixes  
reworded and fixed some grammar to make it easier to read  
before it was perfectly fine, but i decided to fix it anyways  

* Added back the useful functions used by mod makers.  

Added back:  
ProxyColor.RandFromTable()  
ProxyColor.Random()  

* Add files via upload  
* Update README.md  
* Update addon.json  
* Delete ClientsideMenuPreview.gif  

Co-authored-by: NotAKidOnSteam <37721153+NotAKidOnSteam@users.noreply.github.com>